Fix bug in backward for empty masks
